using react icons and moved the appbar to ui/custom
